refactor: modularize postback telemetry engine into lib/audit-postback module
